task 25

+++ app/src/main/java/com/example/bankaccount/BankAccount.java
@@ -44,7 +44,7 @@ public class BankAccount {
         SavingBalance = savingBalance;
     }
 
-    public static double getTotalAmountOfMoney() {
+    public static int getTotalAmountOfMoney() {
         return totalAmountOfMoney;
     }
